[LPN] code restructuring to enable adding LPN
