+++ src/lib/simple_baselines/dataset/coco.py
@@ -19,8 +19,8 @@ import numpy as np
 from pycocotools.coco import COCO
 from pycocotools.cocoeval import COCOeval
 
-from dataset.JointsDataset import JointsDataset
-from nms.nms import oks_nms
+from simple_baselines.dataset.JointsDataset import JointsDataset
+from simple_baselines.nms.nms import oks_nms
 
 
 logger = logging.getLogger(__name__)
