+++ src/lib/simple_baselines/models/__init__.py
@@ -8,4 +8,4 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import models.pose_resnet
+import simple_baselines.models.pose_resnet
